+++ lib/dicey/die_foundry.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
+require_relative "abstract_die"
 require_relative "numeric_die"
 require_relative "regular_die"
+require_relative "static_die"
 
 require_relative "mixins/rational_to_integer"
 
@@ -12,7 +14,8 @@ module Dicey
     include Mixins::RationalToInteger
 
     # Regexp for matching a possible count.
-    PREFIX = /(?:(?<count>[1-9]\d*+)?+d)?+/i
+    PREFIX = '(?:(?<count>[1-9]\d*+)?+[Dd])?+'
+
     # Regexp for an integer number.
     INTEGER = /(?:-?\d++)/
     # Regexp for a (possibly) fractional number.
@@ -27,6 +30,8 @@ module Dicey
       # Integer range goes into the NumericDie mold.
       [/\A#{PREFIX}\(?(?<begin>#{INTEGER})(?:[-–—…]|\.{2,3})(?<end>#{INTEGER})\)?\z/,
        :range_mold].freeze,
+      # Sign-prefixed value goes into the StaticDie mold.
+      [/\A#{PREFIX}\(?(?<sign>\+|-)(?<value>#{STRING})\)?\z/, :static_mold].freeze,
       # List of numbers goes into the NumericDie mold.
       [/\A#{PREFIX}\(?(?<sides>#{INTEGER}(?:(?:,#{INTEGER})++,?+|,))\)?\z/,
        :weirdly_shaped_mold].freeze,
@@ -43,6 +48,7 @@ module Dicey
     # Following definitions are recognized:
     # - positive integer (like "6" or "20"), which produces a {RegularDie};
     # - integer range (like "3-6" or "(-5..5)"), which produces a {NumericDie};
+    # - signed valie (like "+3" or "(-ABC)"), which produces a {StaticDie};
     # - list of integers (like "(3,4,5)", "-1,0,1", or "2,"), which produces a {NumericDie};
     # - list of decimal numbers (like "0.5,0.2,0.8" or "(2.0,)"), which produces a {NumericDie},
     #   but uses +Rational+ for values to maintain precise results;
@@ -82,6 +88,15 @@ module Dicey
       build_dice(NumericDie, definition[:count], first..last)
     end
 
+    def static_mold(definition)
+      value = parse_value(definition[:value])
+      if definition[:sign] != "+"
+        value = (Numeric === value) ? -value : -VectorNumber.new([value])
+      end
+
+      build_dice(StaticDie, definition[:count], value)
+    end
+
     def weirdly_shaped_mold(definition)
       build_dice(NumericDie, definition[:count], definition[:sides].split(",").map(&:to_i))
     end
@@ -93,17 +108,19 @@ module Dicey
 
     def cursed_mold(definition)
       sides = definition[:sides].split(",")
-      sides.map! do |side|
-        case side
-        when /\A#{INTEGER}\z/o
-          side.to_i
-        when /\A#{FRACTION}\z/o
-          rational_to_integer(Rational(side))
-        else
-          side.match(STRING)[:side]
-        end
-      end
+      sides.map! { |side| parse_value(side) }
       build_dice(AbstractDie, definition[:count], sides)
+    end
+
+    def parse_value(side)
+      case side
+      when /\A#{INTEGER}\z/o
+        side.to_i
+      when /\A#{FRACTION}\z/o
+        rational_to_integer(Rational(side))
+      else
+        side.match(STRING)[:side]
+      end
     end
 
     def build_dice(die_class, count, sides)