+++ lib/dicey/static_die.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+require_relative "abstract_die"
+
+module Dicey
+  # Static die has only one side and always returns the same value.
+  # Useful to model constants in dice expressions.
+  class StaticDie < AbstractDie
+    # Die's only value.
+    #
+    # @return [Any]
+    attr_reader :value
+
+    # @param value [Any]
+    def initialize(value)
+      @value = value
+      super([@value].freeze)
+    end
+
+    alias current value
+    alias next value
+    alias roll value
+
+    # Return a string representing the die.
+    #
+    # Static dice are represented with a "+" or "-" followed by the absolute value
+    # (except 0 which doesn't have a sign).
+    #
+    # @return [String]
+    def to_s
+      (!@value.respond_to?(:positive?) || @value.positive?) ? "+#{@value}" : @value.to_s
+    end
+  end
+end